requirements.txt
fastapi
uvicorn
Pillow
PyWavelets
numpy
scipy
ffmpeg-python
python-multipart
requests
streamlit
pytest
httpx
uuid